created authRoutes & moved public routes to its own folder
